LayoutTests/storage/indexeddb/mozilla/event-source-expected.txt

Test IndexedDB's event.target.source in success callbacks

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

indexedDB.open(name)
source = event.target.source;
PASS source is null
db = event.target.result
request = db.setVersion('1')
source = event.target.source;
PASS source is db
Deleted all object stores.
objectStore = db.createObjectStore(objectStoreName, { autoIncrement: true });
request = objectStore.add({});
source = event.target.source;
PASS source is objectStore
PASS successfullyParsed is true

TEST COMPLETE
